pdf-rag-ingestion-pipeline-python/pyproject.toml
[project]
name = "pdf-rag-ingestion-pipeline-python"
version = "0.1.0"
description = "End-to-end Python pipeline: PDFs to Markdown to embeddings to RAG, powered by Nutrient PDF-to-Markdown."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Nutrient", email = "support@nutrient.io" },
]
dependencies = [
  "nutrient-dws>=1.0.0",
  "chromadb>=0.5",
  "openai>=1.40",
  "anthropic>=0.40",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.10",
  "pytest>=8",
  "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "retrieval"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
